# life/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Life Pratics project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/admin/register": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Register a new admin user",
                "operationId": "30fd0eb0742412a292aee650b49e6b69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin user registered successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Login as an admin user",
                "operationId": "cdfade812131da2a04186cbcc471ecb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully logged in"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Logout the admin user",
                "operationId": "51fb0e2ee9af7af87829ee01515d4a88",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/me": {
            "get": {
                "tags": [
                    "Admin Authentication"
                ],
                "summary": "Get the authenticated admin user",
                "operationId": "e3303cd5e5c8624feff64280911da831",
                "responses": {
                    "200": {
                        "description": "Authenticated admin user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medico/register": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Register a new medico",
                "operationId": "166d2798a1583106be364de36ee2935c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "med_crm",
                                    "med_nome",
                                    "med_email",
                                    "med_password"
                                ],
                                "properties": {
                                    "med_crm": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "med_nome": {
                                        "type": "string",
                                        "example": "Dr. João"
                                    },
                                    "med_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "dr.joao@example.com"
                                    },
                                    "med_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Médico registrado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/medico/login": {
            "post": {
                "tags": [
                    "Medico"
                ],
                "summary": "Login for medico",
                "operationId": "7798aae331b0b558fee4d0f079ec0432",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "med_email",
                                    "med_password"
                                ],
                                "properties": {
                                    "med_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "dr.joao@example.com"
                                    },
                                    "med_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/medico/logout": {
            "post": {
                "tags": [
                    "Medico"
                ],
                "summary": "Logout for medico",
                "operationId": "c18d33c473ee091fe7afc561aa19798b",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/medico/me": {
            "get": {
                "tags": [
                    "Medico"
                ],
                "summary": "Get medico details",
                "operationId": "6c575b4d036f3008f9dc93b59f6e013b",
                "responses": {
                    "200": {
                        "description": "Medico details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}